soil sensor ok

+++ src/sensors/soil_sensor.h
@@ -1,7 +1,9 @@
 #ifndef SOIL_SENSOR_H
 #define SOIL_SENSOR_H
 
+#include <stdint.h>
+
 void soil_sensor_init(void);
-int soil_sensor_read(void);
+uint16_t soil_sensor_read(void);
 
 #endif
